backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
transformers==4.46.3
pillow==11.0.0
torch>=2.6.0
torchvision>=0.21.0
pydantic==2.10.2
python-multipart==0.0.9
